Python Automator.
This script allows for recording of all movements with keyboard and mouse with saving and replaying. This allows for any repeating tasks to be recorded and replayed so an individual does not have to perform that repeating task. This also includes all time delays between clicks and keyboard movements for more accurate replaying with less errors.