Automatic NRCCUA ACT information exporter
This script allows for pulling of any information from NRCCUA provided text documents with student ACT information. Script also formats and pulls all the information from the large PDF of student information provided by NRCCUA and pushes and divides it into smaller PDFs named after the students they are assigned to. This automation process minimizes 45 minutes or so a day depending on the amount of ACTS to 2 minutes.